Modify: 全体的に型の整理
