+++ src/pages/api/fetchMicrocmsData.ts
@@ -1,19 +1,6 @@
 import { client } from 'libs/client'
 import { dateToString } from 'src/utils/dateToString'
-import { PostDataType } from 'types/PostDataType'
-import { MicrocmsApi } from 'types/microcmsApi'
-import { MicrocmsData } from 'types/microcmsData'
-
-// ToDo: limitの引数をオプショナルにして、引数あればそのクエリ等を適用、なければ999件取得にする
-
-// type Queries = {
-//   [key: string]: {
-//     offset?: number
-//     limit?: number
-//   }
-// }
-
-// type PartialQueries = Partial<Queries>
+import { PostDataType, MicrocmsData } from 'types/microCms'
 
 async function fetchMicrocmsData(): Promise<PostDataType[]> {
   const microcmsData = await client.getList({
@@ -24,8 +11,12 @@ async function fetchMicrocmsData(): Promise<PostDataType[]> {
   // microCMSのデータ取得
   const microcmsPostData: PostDataType[] = microcmsData.contents.map(
     (item: MicrocmsData) => {
-      const createdAt = dateToString(item.createdAt, 'YYYY/MM/DD')
-      const updatedAt = dateToString(item.updatedAt, 'YYYY/MM/DD')
+      const created_at = dateToString(item.created_at, 'YYYY/MM/DD')
+      const updated_at =
+        item.updated_at === undefined
+          ? ''
+          : dateToString(item.updated_at, 'YYYY/MM/DD')
+
       return {
         id: item.id,
         title: item.title,
@@ -33,8 +24,8 @@ async function fetchMicrocmsData(): Promise<PostDataType[]> {
         description: item.description || null,
         categoryId: item.category.id,
         categoryName: item.category.name,
-        updatedAt,
-        createdAt,
+        created_at,
+        updated_at,
         eyecatch: item.eyecatch.url,
       }
     }
