+++ src/pages/category/[categoryId]/[pageNum].tsx
@@ -20,8 +20,7 @@ import fetchAsideCategory from 'src/pages/api/fetchAsideCategory'
 import { mediaQuery } from 'src/utils/Breakpoints'
 import { dateToString } from 'src/utils/dateToString'
 import { paginationRange } from 'src/utils/paginationRange'
-import { PostDataType } from 'types/PostDataType'
-import { MicrocmsData } from 'types/microcmsData'
+import { PostDataType, MicrocmsData } from 'types/microCms'
 
 export default function CategoryId({
   blog,
@@ -69,8 +68,12 @@ export const getStaticProps: GetStaticProps = async (
   })
 
   const data = microcmsData.contents.map((item: MicrocmsData) => {
-    const createdAt = dateToString(item.createdAt, 'YYYY/MM/DD')
-    const updatedAt = dateToString(item.updatedAt, 'YYYY/MM/DD')
+    const created_at = dateToString(item.created_at, 'YYYY/MM/DD')
+    const updated_at =
+      item.updated_at === undefined
+        ? ''
+        : dateToString(item.updated_at, 'YYYY/MM/DD')
+
     return {
       id: item.id,
       title: item.title,
@@ -78,8 +81,8 @@ export const getStaticProps: GetStaticProps = async (
       description: item.description || null,
       categoryId: item.category.id,
       categoryName: item.category.name,
-      updatedAt,
-      createdAt,
+      updated_at,
+      created_at,
       eyecatch: item.eyecatch.url,
     }
   })
