+++ src/pages/category/zenn/[pageNum].tsx
@@ -14,7 +14,7 @@ import fetchAsideCategory from 'src/pages/api/fetchAsideCategory'
 import fetchZennData from 'src/pages/api/fetchZennData'
 import { mediaQuery } from 'src/utils/Breakpoints'
 import { paginationRange } from 'src/utils/paginationRange'
-import { PostDataType } from 'types/PostDataType'
+import { PostDataType } from 'types/microCms'
 
 export default function CategoryId({
   blog,
