--- types/CategoryCountAndPost.ts
@@ -1,5 +1,0 @@
-export type CategoryCountAndPost = {
-  categoryName: string
-  categoryId: string
-  totalCount: string
-}
