--- types/microCmsPostData.ts
@@ -1,8 +1,0 @@
-export type microCmsPostData = {
-  id: string
-  createdAt: string
-  updatedAt: string
-  publishedAt: string
-  revisedAt: string
-  name: string
-}
