--- types/microcmsApi.ts
@@ -1,9 +1,0 @@
-import { MicrocmsData } from "./microcmsData"
-
-export type MicrocmsApi = {
-  contents: MicrocmsData[]
-  limit: number
-  offset: number
-  totalCount: number
-  id: string
-}
