--- types/storageType.ts
@@ -1,13 +1,0 @@
-export type StorageType = {
-  id: string
-  img: {
-    url: string
-  }
-  title: string
-  tags: {
-    tag: string
-  }[]
-  message: string
-  github: string
-  website: string
-}
